back-end/requirements.txt
aniso8601==9.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-dotenv==0.21.1
pytz==2022.6
requests==2.28.1
six==1.16.0
urllib3==1.26.13
Werkzeug==2.2.2
